+++ src/optimizers/CHICKEN_SWARM_QUANTUM/controller_CHICKEN_QUANTUM.py
@@ -62,7 +62,7 @@ class CONTROLLER_CHICKEN_QUANTUM():
         OUT_VARS = int(optimizerParams['num_output'][0])                   # Number of output variables
         TARGETS = list(optimizerParams['target_values'][0])                # Target values for output
         E_TOL = float(optimizerParams['tolerance'][0])                     # Convergence Tolerance (This is a radius)       
-        MAXIT = float(optimizerParams['max_iterations'][0])                # Maximum allowed iterations 
+        MAXIT = int(optimizerParams['max_iterations'][0])                # Maximum allowed iterations 
         BOUNDARY = int(optimizerParams['boundary'][0])                     # int boundary 1 = random,      2 = reflecting
         RN = int(optimizerParams['rooster_number'][0])                # Target values for output
         HN = int(optimizerParams['hen_number'][0])                  # Variable time-step extinction coefficient
